refactor: 拆分hover状态组件到不同模块
