+++ src/renderer/components/states/hover/index.ts
@@ -1,0 +1,1 @@
+export { HoverContent } from './HoverContent';
